requirements.txt
discord.py
requests
cfscrape
bs4
fuzzywuzzy
youtube-dl